split ssd file
