+++ gluoncv/model_zoo/ssd/__init__.py
@@ -2,4 +2,5 @@
 """Single-shot Object Detection."""
 from __future__ import absolute_import
 from .ssd import *
+from .presets import *
 from .vgg_atrous import *
